Add option "-d" for not running doctor script
installation is always verified by rbenv-doctor which will always trigger an error message: 
```
Checking for `rbenv' in PATH: which: no rbenv in (/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin)
not found
```
And then exit. In order to let the script finish the installation an extra option `-d` is added in order to skip the doctor script. Running doctor still a good idea when your are running it as an user, but not when you are trying to automate it.